@@ -6,6 +6,11 @@
         {
             "name": "Amine Ben hammou",
             "email": "webneat@gmail.com"
+        },
+        {
+            "name": "Wojtek Brozyna",
+            "email": "wojciech.brozyna@gmail.com",
+            "role": "Maintainer (fork)"
         }
     ],
     "autoload": {
@@ -14,9 +19,17 @@
         }
     },
     "require": {
-        "php": "^7.0"
+        "php": "^8.4"
     },
     "require-dev": {
-        "phpunit/phpunit": "^6.3"
+        "phpunit/phpunit": "^10.5",
+        "squizlabs/php_codesniffer": "^3.10",
+        "rector/rector": "^2.2"
+    },
+    "scripts": {
+        "run-tests": "vendor/bin/phpunit --display-deprecations --display-phpunit-deprecations",
+        "code-sniff": "vendor/bin/phpcs -d memory_limit=512M -p src --standard=PSR12",
+        "code-fix": "vendor/bin/phpcbf -d memory_limit=512M -p src --standard=PSR12",
+        "rector": "vendor/bin/rector process"
     }
 }
