@@ -5,20 +5,19 @@ use Tarsana\Filesystem\Adapters\Local;
 use Tarsana\Filesystem\Directory;
 use Tarsana\Filesystem\File;
 
-
-class FileTest extends PHPUnit\Framework\TestCase {
-
+class FileTest extends PHPUnit\Framework\TestCase
+{
     protected $filePath;
 
     protected $file;
 
-    public function setUp()
+    public function setUp(): void
     {
-        $this->filePath = DEMO_DIR.'/temp.txt';
+        $this->filePath = DEMO_DIR . '/temp.txt';
         $this->file = new File($this->filePath);
     }
 
-    public function test_creates_file_if_missing()
+    public function test_creates_file_if_missing(): void
     {
         unlink($this->filePath);
         $this->assertFalse(is_file($this->filePath));
@@ -27,20 +26,18 @@ class FileTest extends PHPUnit\Framework\TestCase {
         $this->assertTrue(is_file($this->filePath));
     }
 
-    /**
-     * @expectedException Tarsana\Filesystem\Exceptions\FilesystemException
-     */
-    public function test_throws_exception_when_directory_exists_with_same_path()
+    public function test_throws_exception_when_directory_exists_with_same_path(): void
     {
-        $file = new File(DEMO_DIR.'/folder1');
+        $this->expectException(\Tarsana\Filesystem\Exceptions\FilesystemException::class);
+        $file = new File(DEMO_DIR . '/folder1');
     }
 
-    public function test_gets_filesystem_instance()
+    public function test_gets_filesystem_instance(): void
     {
         $this->assertTrue($this->file->fs() instanceof Filesystem);
     }
 
-    public function test_exists()
+    public function test_exists(): void
     {
         $this->assertTrue($this->file->exists());
 
@@ -48,37 +45,35 @@ class FileTest extends PHPUnit\Framework\TestCase {
         $this->assertFalse($this->file->exists());
     }
 
-    public function test_gets_and_sets_absolute_path()
+    public function test_gets_and_sets_absolute_path(): void
     {
         $this->assertEquals($this->filePath, $this->file->path());
 
-        $this->file->path(DEMO_DIR.'/folder2/temp.txt', true);
-        $this->assertEquals(DEMO_DIR.'/folder2/temp.txt', $this->file->path());
-        $this->assertTrue(is_file(DEMO_DIR.'/folder2/temp.txt'));
+        $this->file->path(DEMO_DIR . '/folder2/temp.txt', true);
+        $this->assertEquals(DEMO_DIR . '/folder2/temp.txt', $this->file->path());
+        $this->assertTrue(is_file(DEMO_DIR . '/folder2/temp.txt'));
 
         $this->file->path($this->filePath);
         $this->assertEquals($this->filePath, $this->file->path());
         $this->assertTrue(is_file($this->filePath));
     }
 
-    /**
-     * @expectedException Tarsana\Filesystem\Exceptions\FilesystemException
-     */
-    public function test_throws_exception_when_already_exists_path()
+    public function test_throws_exception_when_already_exists_path(): void
     {
-        $path = DEMO_DIR.'/temp-2.txt';
+        $this->expectException(\Tarsana\Filesystem\Exceptions\FilesystemException::class);
+        $path = DEMO_DIR . '/temp-2.txt';
         file_put_contents($path, '');
         $this->file->path($path);
     }
 
-    public function test_gets_and_sets_name()
+    public function test_gets_and_sets_name(): void
     {
         $this->assertEquals('temp.txt', $this->file->name());
 
         $this->file->name('new-temp.txt');
         $this->assertEquals('new-temp.txt', $this->file->name());
-        $this->assertEquals(DEMO_DIR.'/new-temp.txt', $this->file->path());
-        $this->assertTrue(is_file(DEMO_DIR.'/new-temp.txt'));
+        $this->assertEquals(DEMO_DIR . '/new-temp.txt', $this->file->path());
+        $this->assertTrue(is_file(DEMO_DIR . '/new-temp.txt'));
 
         $this->file->name('temp.txt');
         $this->assertEquals('temp.txt', $this->file->name());
@@ -86,15 +81,13 @@ class FileTest extends PHPUnit\Framework\TestCase {
         $this->assertTrue(is_file($this->filePath));
     }
 
-    /**
-     * @expectedException Tarsana\Filesystem\Exceptions\FilesystemException
-     */
-    public function test_throws_exception_when_invalid_name()
+    public function test_throws_exception_when_invalid_name(): void
     {
+        $this->expectException(\Tarsana\Filesystem\Exceptions\FilesystemException::class);
         $this->file->name('');
     }
 
-    public function test_gets_and_sets_permissions()
+    public function test_gets_and_sets_permissions(): void
     {
         chmod($this->filePath, 0755);
         $this->assertEquals('0755', $this->file->perms());
@@ -103,7 +96,7 @@ class FileTest extends PHPUnit\Framework\TestCase {
         $this->assertEquals('0777', $this->file->perms());
     }
 
-    public function test_is_writable()
+    public function test_is_writable(): void
     {
         chmod($this->filePath, 0444);
         $this->assertFalse($this->file->isWritable());
@@ -112,7 +105,7 @@ class FileTest extends PHPUnit\Framework\TestCase {
         $this->assertTrue($this->file->isWritable());
     }
 
-    public function test_is_executable()
+    public function test_is_executable(): void
     {
         chmod($this->filePath, 0444);
         $this->assertFalse($this->file->isExecutable());
@@ -123,7 +116,7 @@ class FileTest extends PHPUnit\Framework\TestCase {
         chmod($this->filePath, 0777);
     }
 
-    public function test_sets_and_gets_extension()
+    public function test_sets_and_gets_extension(): void
     {
         $this->assertEquals('txt', $this->file->extension());
 
@@ -139,7 +132,7 @@ class FileTest extends PHPUnit\Framework\TestCase {
         $this->file->name('temp.txt');
     }
 
-    public function test_gets_and_sets_content()
+    public function test_gets_and_sets_content(): void
     {
         file_put_contents($this->filePath, 'The obligatory Hello World !');
         $this->assertEquals('The obligatory Hello World !', $this->file->content());
@@ -148,7 +141,7 @@ class FileTest extends PHPUnit\Framework\TestCase {
         $this->assertEquals('A new content', file_get_contents($this->filePath));
     }
 
-    public function test_append()
+    public function test_append(): void
     {
         file_put_contents($this->filePath, 'The obligatory ');
         $this->file->append('Hello World !');
@@ -156,26 +149,26 @@ class FileTest extends PHPUnit\Framework\TestCase {
         $this->assertEquals('The obligatory Hello World !', file_get_contents($this->filePath));
     }
 
-    public function test_copy_as()
+    public function test_copy_as(): void
     {
         file_put_contents($this->filePath, 'Some content');
-        $copy = $this->file->copyAs(DEMO_DIR.'/copies/new-temp.txt');
+        $copy = $this->file->copyAs(DEMO_DIR . '/copies/new-temp.txt');
 
-        $this->assertTrue(is_file(DEMO_DIR.'/copies/new-temp.txt'));
+        $this->assertTrue(is_file(DEMO_DIR . '/copies/new-temp.txt'));
         $this->assertTrue($copy instanceof File);
-        $this->assertEquals(DEMO_DIR.'/copies/new-temp.txt', $copy->path());
-        $this->assertEquals('Some content', file_get_contents(DEMO_DIR.'/copies/new-temp.txt'));
+        $this->assertEquals(DEMO_DIR . '/copies/new-temp.txt', $copy->path());
+        $this->assertEquals('Some content', file_get_contents(DEMO_DIR . '/copies/new-temp.txt'));
 
-        (new Directory(DEMO_DIR.'/copies'))->remove();
+        new Directory(DEMO_DIR . '/copies')->remove();
     }
 
-    public function test_gets_hash()
+    public function test_gets_hash(): void
     {
         file_put_contents($this->filePath, 'The obligatory Hello World !');
         $this->assertEquals(md5_file($this->filePath), $this->file->hash());
     }
 
-    public function test_remove()
+    public function test_remove(): void
     {
         $file = new File($this->filePath);
         $this->assertTrue(
@@ -187,10 +180,9 @@ class FileTest extends PHPUnit\Framework\TestCase {
         $this->assertFalse(file_exists($this->filePath));
     }
 
-    public function tearDown()
+    public function tearDown(): void
     {
-        remove(DEMO_DIR.'/temp.txt');
-        remove(DEMO_DIR.'/temp-2.txt');
+        remove(DEMO_DIR . '/temp.txt');
+        remove(DEMO_DIR . '/temp-2.txt');
     }
-
 }
